refactor: simplify boolean assertions in tests
Replaced `assert_eq!(state, false)` with `assert!(!state)` for clearer and more concise boolean assertions.

refactor: simplify match result handling

Removed redundant explicit `return` statements in the match branches for `Self::detect_with_get_version_ex()`, simplifying the code.

refactor: initialize structs with default

Initialized `OSVERSIONINFOEXW` and `PROCESS_MEMORY_COUNTERS` structs with `Default::default()` when setting a specific field, improving readability and consistency.

refactor: remove redundant `as u32` cast

Removed the unnecessary `as u32` cast for `BI_RGB.0` as `BI_RGB.0` is already of type `u32`.

+++ src/utils/icon_extractor.rs
@@ -234,7 +234,7 @@ fn hicon_to_rgba_bytes(hicon: HICON) -> Result<Vec<u8>> {
         bmi.bmiHeader.biHeight = -(height as i32); // Negative for top-down DIB
         bmi.bmiHeader.biPlanes = 1;
         bmi.bmiHeader.biBitCount = 32;
-        bmi.bmiHeader.biCompression = BI_RGB.0 as u32;
+        bmi.bmiHeader.biCompression = BI_RGB.0;
 
         // Allocate buffer for bitmap data (BGRA format)
         let mut buffer = vec![0u8; width * height * 4];
